Select new mount by default if requested
Summary:
By default on open we select the first mount. This matches Networks and
User Accounts.

Some distros place a .desktop file with the name "New Share" that opens
this KCM. Having it preselect an entry at this point is confusing.

Changing the Exec line to
"Exec=kcmshell5 sambamount --args --new" will default to selecting the
"New Share" entry. The rest is unchanged.

Reviewers: adridg, chinmoyr, apol

Reviewed By: apol

Differential Revision: https://phabricator.kde.org/D17964